+++ Src/read_options.F90
@@ -1386,11 +1386,13 @@ subroutine read_options( na, ns, nspin )
 
   ! Target Temperature and Pressure
   tt = fdf_get('MD.TargetTemperature',0.0_dp,'K')
-  tp = fdf_get('MD.TargetPressure',0.0_dp,'Ry/Bohr**3')
-  !
-  ! Used for now for the call of the PR md routine if quenching
-  if (idyn == 3 .AND. iquench > 0) call set_target_stress()
 
+  call fdf_deprecated("MD.TargetPressure", "Target.Pressure")
+  tp = fdf_get('MD.TargetPressure',0.0_dp,'Ry/Bohr**3')
+  tp = fdf_get('Target.Pressure',tp,'Ry/Bohr**3')
+
+  ! Used for now for the call of the PR md routine if quenching
+  if (idyn == 3 .AND. iquench > 0) call set_target_stress(tp)
 
   ! Mass of Nose variable
   mn = fdf_get('MD.NoseMass',100._dp,'Ry*fs**2')
@@ -1400,10 +1402,10 @@ subroutine read_options( na, ns, nspin )
 
   if (idyn==2 .or. idyn==4) then
      if (ionode) then
-        write(6,6) 'redata: Nose mass',mn/eV,' eV/fs**2'
+        write(6,6) 'redata: Nose mass',mn/eV,' eV*fs**2'
      endif
      if (cml_p) then
-        call cmlAddParameter( xf    = mainXML,              &
+        call cmlAddParameter( xf    = mainXML,&
              name  = 'MD.NoseMass',        &
              value = mn,                   &
              units = 'siestaUnits:Ry_fs__2')
@@ -1412,12 +1414,12 @@ subroutine read_options( na, ns, nspin )
 
   if (idyn==3 .or. idyn==4) then
      if (ionode) then
-        write(6,6) 'redata: Parrinello-Rahman mass',mpr/eV,' eV/fs**2'
+        write(6,6) 'redata: Parrinello-Rahman mass',mpr/eV,' eV*fs**2'
      endif
      if (cml_p) then
         call cmlAddParameter( xf    = mainXML,                   &
              name  = 'MD.ParrinelloRahmanMass', &
-             value = mn,                        &
+             value = mpr, &
              units = 'siestaUnits:Ry_fs__2' )
      endif
   endif